- added sound-support information for HP-UX, - added a QString, which GetInfo_XYZ-functions   can predefine to show the user an error.
svn path=/trunk/kdebase/kcontrol/; revision=20817

+++ kcontrol/info/memory.cpp
@@ -76,7 +76,7 @@ QString format(unsigned long value)
   QString  text;
   double   mb = value / 1048576.0;
   
-  text.sprintf(i18n("%10ld bytes  = %8.2f MB"), value, mb);
+  text.sprintf(i18n("%10u bytes  = %8.2f MB"), ((unsigned int)value), mb);
   return text;
 }
